+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - fixed: (Hedera) Fix missing transactions by specifying query limit
 - fixed: Updated FEVM default gas price settings.
 - fixed: Include deprecated `defaultSettings` field in FEVM's EdgeCurrencyInfo for legacy support.
+- fixed: (Optimism/Sepolia/Holesky) Remove defunct etherscan server URIs.
+- fixed: Broken response cleaner for etherscan v2 proxy module request (used for broadcasts).
 
 ## 4.48.1 (2025-05-27)